Implement templatable agent helm values
This will allow for more flexible configuration cluster-by-cluster
